etl/requirements.txt



dagit==0.14.3
dagster-postgres==0.14.3
dagster==0.14.3
ipython-genutils==0.2.0



pandas
requests
dagster-shell
pydantic>=1.8.0,<2.0.0
sqlalchemy

python-decouple

wget

tqdm
grpcio==1.46.3
grpcio-health-checking==1.43.0
protobuf<3.21.0


beautifulsoup4==4.11.1
lxml==4.9.0


sqlalchemy_utils
